value uses rc and copy on write to do less cloning

+++ src/main.rs
@@ -1,4 +1,5 @@
 #![feature(type_alias_impl_trait)]
+#![feature(arc_unwrap_or_clone)]
 #![feature(once_cell)]
 use std::{
     fs,
@@ -96,5 +97,5 @@ fn start() -> Result<i32, ShellErrorKind> {
         },
     };
 
-    Ok(num_traits::clamp(status, i32::MIN as i128, i32::MAX as i128) as i32)
+    Ok(num_traits::clamp(status, i32::MIN as i64, i32::MAX as i64) as i32)
 }
